+++ docs/whats_new.rst
@@ -52,6 +52,7 @@ Enhancements
 - Ensuring consistency on the expose of the activation function (:gh:`637` by `Bruno Aristimunha`_ )
 - Adding :class:`braindecode.models.SyncNet` (:gh:`643` by `Bruno Aristimunha`_ )
 - Including PytorchAudio as dependency and remove copied code (:gh:`661` by `Bruno Aristimunha`_)
+- Adding :class:`braindecode.models.EEGSym` (:gh:`664` by `Bruno Aristimunha`_ )
 
 
 Bugs